Make audio/image pages without ids show a 404
