+++ src/pages/image/index.vue
@@ -1,0 +1,11 @@
+<template><VFourOhFour /></template>
+
+<script>
+import { defineComponent } from '@vue/composition-api'
+import VFourOhFour from '~/components/VFourOhFour.vue'
+
+export default defineComponent({
+  components: { VFourOhFour },
+  layout: 'blank',
+})
+</script>
